refactor: user service interface 생성
